refactor: 헥사고날 에러코드의 종속성 제거 및 추상화
