+++ src/main/java/com/agenticcp/core/domain/cloud/exception/CloudErrorCode.java
@@ -6,15 +6,15 @@ import org.springframework.http.HttpStatus;
 
 /**
  * 클라우드 도메인 일반 에러 코드
- * (S3, AWS 등 특정 서비스별 에러는 각각의 ErrorCode 클래스 참조)
+ * (특정 CSP 구현 어댑터는 이 코드를 사용해 도메인으로 예외를 전달합니다)
  */
 public enum CloudErrorCode implements BaseErrorCode {
     
     // 4000-4999: 클라우드 도메인
     CAPABILITY_NOT_DEFINED(HttpStatus.BAD_REQUEST, 4001, "지원 Capability가 정의되지 않았습니다."),
     UNSUPPORTED_OPERATION(HttpStatus.BAD_REQUEST, 4002, "지원되지 않는 작업입니다."),
-    PROVIDER_UNAVAILABLE(HttpStatus.BAD_GATEWAY, 4003, "클라우드 제공자 서비스와 통신할 수 없습니다."),
-    RATE_LIMIT_EXCEEDED(HttpStatus.TOO_MANY_REQUESTS, 4004, "호출 제한을 초과했습니다."),
+    CLOUD_PROVIDER_UNAVAILABLE(HttpStatus.BAD_GATEWAY, 4003, "클라우드 제공자 서비스와 통신할 수 없습니다."),
+    API_RATE_LIMIT_EXCEEDED(HttpStatus.TOO_MANY_REQUESTS, 4004, "API 호출 제한을 초과했습니다."),
     API_TIMEOUT(HttpStatus.GATEWAY_TIMEOUT, 4005, "클라우드 API 호출이 시간 초과되었습니다."),
     MAPPING_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, 4006, "CSP 응답 매핑에 실패했습니다."),
     
@@ -37,9 +37,14 @@ public enum CloudErrorCode implements BaseErrorCode {
     // 일반적인 클라우드 에러
     CLOUD_CONNECTION_FAILED(HttpStatus.SERVICE_UNAVAILABLE, 4030, "클라우드 서비스 연결에 실패했습니다."),
     CLOUD_OPERATION_TIMEOUT(HttpStatus.REQUEST_TIMEOUT, 4031, "클라우드 작업 시간이 초과되었습니다."),
-    CLOUD_RESOURCE_LIMIT_EXCEEDED(HttpStatus.TOO_MANY_REQUESTS, 4032, "클라우드 리소스 한도를 초과했습니다."),
+    RESOURCE_QUOTA_EXCEEDED(HttpStatus.TOO_MANY_REQUESTS, 4032, "클라우드 리소스 할당량을 초과했습니다."),
     CLOUD_TAG_OPERATION_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, 4033, "클라우드 리소스 태그 작업이 실패했습니다."),
-    CLOUD_METADATA_SERIALIZATION_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, 4034, "클라우드 메타데이터 직렬화가 실패했습니다.");
+    CLOUD_METADATA_SERIALIZATION_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, 4034, "클라우드 메타데이터 직렬화가 실패했습니다."),
+    PERMISSION_DENIED(HttpStatus.FORBIDDEN, 4035, "클라우드 리소스에 대한 권한이 없습니다."),
+    INVALID_REGION(HttpStatus.BAD_REQUEST, 4036, "유효하지 않은 클라우드 리전입니다."),
+    INVALID_REQUEST(HttpStatus.BAD_REQUEST, 4037, "유효하지 않은 클라우드 API 요청입니다."),
+    CLOUD_NETWORK_ERROR(HttpStatus.SERVICE_UNAVAILABLE, 4038, "클라우드 네트워크 통신에 실패했습니다."),
+    CLOUD_CONFIGURATION_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, 4039, "클라우드 설정이 잘못되었습니다.");
     
     private final HttpStatus httpStatus;
     private final int codeNumber;
